// frontend/src/app/models/task-info.ts
export class TaskInfo {
  id?: number;
  projectName?: string;
  dueDate?: Date;
  startDate?: Date;
  assignTo?: string;
  assignedDate?: Date;
  projectPriority?: any;
  department?: string;
  departments?: any;
  projectTitle?: string;
  projectDescription?: string;
  attachment?: any;
  status?: string;
  assignedBy?: string;
  assignedTo?: string;
  title?: string;
  priority?: string;
  description?: string;
}